requirements.txt
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
tqdm>=4.66.0